cmake build of glog. fix some warnings. fix bug in << task *

+++ glog/demangle.cc
@@ -353,7 +353,7 @@ static void MaybeCancelLastSeparator(State *state) {
 // namespace.
 static bool IdentifierIsAnonymousNamespace(State *state) {
   const char anon_prefix[] = "_GLOBAL__N_";
-  return (state->number > sizeof(anon_prefix) - 1 &&  // Should be longer.
+  return (state->number > (int)sizeof(anon_prefix) - 1 &&  // Should be longer.
           StrPrefix(state->mangled_cur, anon_prefix));
 }
